Fix comp warning

+++ src/deps/mpr/mprLib.c
@@ -1783,7 +1783,7 @@ static void printTracking()
             for (np = &lp->names[0]; *np && np < &lp->names[MPR_TRACK_NAMES]; np++) {
                 if (*np) {
                     if (np == lp->names) {
-                        printf("%10ld %-24s\n", lp->count, *np);
+                        printf("%10d %-24s\n", (int) lp->count, *np);
                     } else {
                         printf("           %-24s\n", *np);
                     }
